Update README.md
Imagine you buy a chocolate bar and inside, there is a code for a chance to win a free cinema voucher. You go to the website, enter the code and the website either gives you a voucher number, or returns to the code entry screen. My code simulates this process. It uses classes to make it neat and readable. It contains a list of winning codes, and checks the user entry against the list to determine if the code is a winner. It uses time.sleep() if too many incorrect entries are made: this is to deter spam/bots.